requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
attrs==19.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
Flask==1.0.3
future==0.17.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pathspec==0.7.0
pycparser==2.19
python-telegram-bot==11.0.0
regex==2020.1.8
requests==2.22.0
six==1.12.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.2
Werkzeug==0.15.4
